Bug: commiting to save all my attempts at fixing this issue. Currently broken.
