+++ client/package.json
@@ -9,6 +9,7 @@
     "rc-slider": "^8.6.3",
     "react": "^16.5.2",
     "react-dom": "^16.5.2",
+    "react-input-range": "^1.3.0",
     "react-leaflet": "^2.1.0",
     "react-modal": "^3.6.1",
     "react-router-dom": "^4.3.1",
