+++ client/src/Components/BorderBay/BorderBay.css
@@ -11,6 +11,10 @@
   border: 1px solid black;
 }
 .Country_Border .Country_Border-Slider {
-  width: 100%;
-  height: 60px;
+  position: relative;
+  z-index: 10;
+}
+.Country_Border .Country_Border-Slider .Slider1 {
+  position: relative;
+  z-index: 100;
 }
