+++ client/src/Components/BorderBay/BorderBay.less
@@ -11,9 +11,13 @@
     width: 50%;
     border: 1px solid black;
   }
-
   .Country_Border-Slider {
-    width: 100%;
-    height: 60px;
+    position: relative;
+    z-index: 10;
+
+    .Slider1 {
+      position: relative;
+      z-index: 100;
+    }
   }
 }
